BACKEND/waste_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
africastalking
python-decouple
daphne
requests
urllib3<2.0.0
